Separação do stepdefs
